scrap-malt/requirements.txt
fake-useragent==0.1.11
pymongo==3.12.0
selenium==3.141.0
urllib3==1.26.6
google.cloud==0.34.0